scraping-tools/pyproject.toml
[tool.poetry]
name = "scraping-tools"
version = "0.1.0"
description = ""
authors = ["Christopher Bartling <chris.bartling@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.1.1"
requests = "^2.31.0"
pandas = "^2.2.1"
beautifulsoup4 = "^4.12.3"
python-dotenv = "^1.0.1"
mechanize = "^0.4.9"
playwright = "^1.42.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
